vim: Make auto shfmt respect .editorconfig
The codefmt plugin calls shfmt passing the source code through stdin.
When this happens, shfmt doesn't know where the file lives and therefore
doesn't find .editorconfig.

shfmt's solution to this is the --filename flag, which tells shfmt which
file is being formatted even if the content itself is passed via stdin.

To make the codefmt Vim plugin use that flag, a new function that
returns --filename <path/to/file/being/edited> is created, and
configured via Glaive as the generator of cmdline options that are
passed to shfmt every time it's called.